pyproject.toml
[tool.poetry]
name = "cslproject"
version = "0.1.0"
description = ""
authors = ["amgookool <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "^1.4.15"
openpyxl = "^3.0.7"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
